Reworked SerializedRow and EnhancedValue classes
We have reworked the mentioned classes to make it easier to write new
conversions from types such as boost::gregorian::date to
mysqlx::EnhancedValue. Through SerializedRow, we gain the ability to
convert from any object T (which's attributes are constrained to
convertible types to EnhancedValue) to SerializedRow and subsequently to
mysqlx::Row.